requirements.txt
torch
torchvision
segmentation-models-pytorch
nibabel
pydicom
Pillow
numpy
tqdm
